rename delete command
