+++ src/main/java/seedu/address/logic/parser/statistics/PlotProductSalesCommandParser.java
@@ -6,7 +6,7 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_END_DATE;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_START_DATE;
 
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.statistics.PlotProductSalesCommand;
+import seedu.address.logic.commands.product.PlotProductSalesCommand;
 import seedu.address.logic.parser.ArgumentMultimap;
 import seedu.address.logic.parser.ArgumentTokenizer;
 import seedu.address.logic.parser.Parser;
