+++ src/main/java/seedu/address/logic/parser/transaction/DeleteTransactionCommandParser.java
@@ -3,7 +3,7 @@ package seedu.address.logic.parser.transaction;
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.transaction.DeleteTransactionCommand;
+import seedu.address.logic.commands.transaction.UndoTransactionCommand;
 import seedu.address.logic.parser.Parser;
 import seedu.address.logic.parser.ParserUtil;
 import seedu.address.logic.parser.exceptions.ParseException;
@@ -11,20 +11,20 @@ import seedu.address.logic.parser.exceptions.ParseException;
 /**
  * Parses input arguments and creates a new DeleteTransactionCommand object
  */
-public class DeleteTransactionCommandParser implements Parser<DeleteTransactionCommand> {
+public class DeleteTransactionCommandParser implements Parser<UndoTransactionCommand> {
 
     /**
      * Parses the given {@code String} of arguments in the context of the DeleteTransactionCommand
      * and returns a DeleteTransactionCommand object for execution.
      * @throws ParseException if the user input does not conform the expected format
      */
-    public DeleteTransactionCommand parse(String args) throws ParseException {
+    public UndoTransactionCommand parse(String args) throws ParseException {
         try {
             Index index = ParserUtil.parseIndex(args);
-            return new DeleteTransactionCommand(index);
+            return new UndoTransactionCommand(index);
         } catch (ParseException pe) {
             throw new ParseException(
-                    String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteTransactionCommand.MESSAGE_USAGE), pe);
+                    String.format(MESSAGE_INVALID_COMMAND_FORMAT, UndoTransactionCommand.MESSAGE_USAGE), pe);
         }
     }
